pyproject.toml
[project]
name = "cyber"
version = "0.1.0"
description = "A simplistic coding editor"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi[standard]>=0.120.2",
    "pydantic-settings>=2.11.0",
    "starlette>=0.49.1",
    "sqlalchemy[asyncio]>=2.0.31",
    "alembic>=1.13.2",
    "llama-index-workflows>=2.8.3",
    "llama-index-core>=0.14.5",
    "llama-index-llms-anthropic>=0.10.0",
    "llama-index-llms-openai>=0.6.6",
    "async-lru>=2.0.5",
    "grep-ast>=0.9.0",
    "scipy>=1.13.0",
    "openinference-instrumentation-llama-index>=4.3.9",
    "opentelemetry-sdk>=1.38.0",
    "opentelemetry-exporter-otlp>=1.38.0",
    "opentelemetry-proto>=1.12.0",
    "arize-phoenix>=12.18.0",
    "genai-prices>=0.0.47",
    "aiofiles>=25.1.0",
    "llama-index-llms-google-genai>=0.8.3",
]

[dependency-groups]
dev = [
    "pytest>=7.4.3,<8.0.0",
    "black>=25.12.0",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
    "pytest-mock>=3.15.1",
    "pytest-asyncio>=0.23.8",
    "pytest-socket>=0.7.0"
]

[tool.uv]
default-groups = ["dev"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py313"
exclude = ["alembic"]

[tool.ruff.format]
exclude = ["alembic"]

[tool.black]
line-length = 88
target-version = ["py313"]
extend-exclude = "alembic"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501",  # line too long, handled by Black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
    "ARG001", # FastAPI routes often receive Request for templating/context
    "T201",   # Allow prints in seeds/examples
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.coverage.run]
source = ["app"]
dynamic_context = "test_function"

[tool.coverage.report]
show_missing = true
sort = "-Cover"

[tool.coverage.html]
show_contexts = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = [
    "--disable-socket",
    "--allow-unix-socket",
]